Refactor grpc tests to not use the same state
